Small fix in layer test

+++ tests/layer_tests/pytorch_tests/test_index_put_.py
@@ -34,9 +34,9 @@ class TestIndexPut_SingleIndices(PytorchLayerTest):
                                              'values': np.array(11).astype(np.float32)},
                                              {'input_tensor': np.random.randn(3, 3).astype(np.float32),
                                              'values': np.array([10, 11, 12]).astype(np.float32)}))
-    @pytest.mark.parametrize('indices', (torch.tensor([0], dtype=torch.int32),
-                                         torch.tensor([1, 2], dtype=torch.int32),
-                                         torch.tensor([0, 1, 2], dtype=torch.int32)))
+    @pytest.mark.parametrize('indices', (torch.tensor([0], dtype=torch.long),
+                                         torch.tensor([1, 2], dtype=torch.long),
+                                         torch.tensor([0, 1, 2], dtype=torch.long)))
     @pytest.mark.parametrize('accumulate', (True, False))
     @pytest.mark.nightly
     @pytest.mark.precommit
@@ -71,12 +71,12 @@ class TestIndexPut_ManyIndices(PytorchLayerTest):
                                              'values': np.array(12).astype(np.float32)},
                                              {'input_tensor': np.random.randn(3, 3, 3).astype(np.float32),
                                              'values': np.array([10, 11, 12]).astype(np.float32)},))
-    @pytest.mark.parametrize('indices', ((torch.tensor([0], dtype=torch.int32),
-                                          torch.tensor([2], dtype=torch.int32)),
-                                         (torch.tensor([1, 2], dtype=torch.int32),
-                                          torch.tensor([0, 1], dtype=torch.int32)),
-                                         (torch.tensor([0, 1], dtype=torch.int32),
-                                          torch.tensor([0, 1], dtype=torch.int32))))
+    @pytest.mark.parametrize('indices', ((torch.tensor([0], dtype=torch.long),
+                                          torch.tensor([2], dtype=torch.long)),
+                                         (torch.tensor([1, 2], dtype=torch.long),
+                                          torch.tensor([0, 1], dtype=torch.long)),
+                                         (torch.tensor([0, 1], dtype=torch.long),
+                                          torch.tensor([0, 1], dtype=torch.long))))
     @pytest.mark.parametrize('accumulate', (True, False))
     @pytest.mark.nightly
     @pytest.mark.precommit
